refactor: ♻️ rename GroupList to GroupsCarousel
