+++ src/lib/components/summary/GroupsCarousel.svelte
